Switched from uint* to uint

+++ cmd/internal/interpolator.go
@@ -10,7 +10,7 @@ import (
 )
 
 func NewInterpolator(problem *Problem, target Target) (*adhier.Interpolator, error) {
-	ic, oc := uint16(target.Inputs()), uint16(target.Outputs())
+	ic, oc := target.Inputs(), target.Outputs()
 
 	var grid adhier.Grid
 	var basis adhier.Basis
